refactor: split device info modules
